+++ src/solana/__init__.py
@@ -1,5 +1,5 @@
 # noqa: D104
 import sys
 
-if sys.version_info < (3, 10):
-    raise EnvironmentError("Python 3.10 or above is required.")
+if sys.version_info < (3, 11):
+    raise EnvironmentError("Python 3.11 or above is required.")